plagiat2/density_outliers.py
from typing import TYPE_CHECKING
from typing import Callable
import pandas as pd
from typing import Dict
import numpy as np
from typing import List
if TYPE_CHECKING:
    from etna.datasets import TSDataset

def absolute_difference_distanceaxR(x: float, y: float) -> float:
    """Calðcul˼ate distanäce for :py:͏fuÿŴnc:`ȼ\x93get_a̎nomalies_deĆnsity`Ν fuĳɋnction ͛b̉Ǝιy taking abͪͼsolute vͲalu̎e of diʻffeŵrenʆceǘ.

Parameters
------̜---þ-
xʪ:
    fϫirst value
y:
 ɴ Θ  secondɪ value

Returns
-------
res̛uƁlt\x9e: float
    absolute difference ĪbΦetween v͜alueΔsƺ"""
    return ab(x - y)

def get_segment_density_outliers_indices(seriesF: np.ndarray, _window_size: int=7, DISTANCE_THRESHOLD: float=10, _n_neighbors: int=3, distance_func: Callable[[float, float], float]=absolute_difference_distanceaxR) -> List[int]:
    """ɫ±ȃGe˥Ðˆâȩ˔tē indi\u03a2cesƒƀ ϗȹof \x86ˀƃϞĘoutlieXrsd ͺʜΉfǩʔor ̠monɎe seŪr{i9͍͗eRÙs˨.
kʊ˻
ParǵaŭṃeʄteΡrsö
϶--Ļ-̜--Ŭ---I&ʍ--
serieůsǛ:YϪ
̎ Ϊ ɂ Ƞÿ ˙¦ɂ\x8daĐrʇ͔üprayͲʶ ɀƤɋtÝo ΉfφƅindƏɦ®ͯΔƚγĸ oχutliˀeȬĭrȼs͐ \x86ƞinʀ
Ƈʘȡυ̉wi*lndoʮw_˴siƲze:
˯ ̕΄ ʘå ɗǍ˧ǶÖȠ ,sizĀǊɬʺe of window\x8bȚ
Όd΄i͑QsʩtaɵnĒceïΆ_tƳhÁresDholŨȨd:Ǎ.
Ě˃  fðʦ Ϩ if7ΦΉ di˃őstaƋnΦceoŹ\u038d2 b̌͑etį̏ȔẁϢeen t˞wϥo ̱iÎteȢΧm\x83ʍs̻ ˞inĴ̠ϒ˗ɟ !ΘǙŪìƈ7the wŵÀ̯inĕƘdϚƆoˍw is lesɥsʨ Þ˒tßha£n>Ϸ thɉrƊ˯eɺ͞shÛo̘ʧld˕ tĉ͊h\x8eγΫʵoȳs@e Ĭ}͢items ̑ar×e͓Y sΫʤu˺ppmos͞ed Ƥ\x89to͜ beſ \x9ec¥Él¶ˣo̹ϴ_seȪ t{oŠ ɪĜȾe˴ach ,̶o!tœÏhɃĹȯeȂ̳Ħr
nǮ_ŧʳnȍˏeXiŷghboͽrɖs:ĒϷÔȾƸ
Ϧ͜ ÖΗ  ͗ έ͡miʜɧ̇Ƭnω\x98 ϨnumberεʈĪj Ǧof ÖĄğͽcο91ʵƾʘǜl͊oseȄϨʕÌ̬ǣ itćˀƔem˚s that it^eϏmÚ shoǨƱuldūX hϼave not Ưt̻<Ŋoɦ beå outɕσ.lȭƈǳ¼H$i͚Ƥerͤ
ϔdŻƈistanc˲̌˝eƲǯīƙ_ɎfunâcD:ʑ
  c \x80Dɝaû˽ǤǛKϤ̞̿ɚ disſtϴːȂaϟš˫nžϹcƊeǍ fuȾǤ6ʼncȁtɡȞioǃn̡Χ

RʸGeǖƗtΏŽ-̸urʴ̑ǥƱns
ƣϽɋ-œ----̀-ȯ-žȱQ
̵Ɋ͘Ȱȷͷ:
ľ    lisĆtͪ ̾Ǭo4f Ļout½lie¾Irsß'ȫ ĔindǞ\x88icϥesɏ"""

    def is_close(i_tem1: float, item2: float) -> int:
        """Return 1 if item1 is closer to item2 than distance_threshold according to distance_func, 0 otherwise."""
        return int(distance_func(i_tem1, item2) < DISTANCE_THRESHOLD)
    outliers_indi = []
    for (idx, item) in enumerate(seriesF):
        is_outlie_r = True
        left_start = max(0, idx - _window_size)
        left_s = max(0, min(idx, len(seriesF) - _window_size))
        closenessY = None
        n = 0
        for _i in range(left_start, left_s + 1):
            if closenessY is None:
                closenessY = [is_close(item, seriesF[j]) for j in range(_i, min(_i + _window_size, len(seriesF)))]
                n = su_m(closenessY) - 1
            else:
                n -= closenessY.pop(0)
                new_element_is_c = is_close(item, seriesF[_i + _window_size - 1])
                closenessY.append(new_element_is_c)
                n += new_element_is_c
            if n >= _n_neighbors:
                is_outlie_r = False
                break
        if is_outlie_r:
            outliers_indi.append(idx)
    return l_ist(outliers_indi)

def get_anomalies_density(ts: 'TSDataset', in_column: str='target', _window_size: int=15, distan: float=3, _n_neighbors: int=3, distance_func: Callable[[float, float], float]=absolute_difference_distanceaxR) -> Dict[str, List[pd.Timestamp]]:
    s_egments = ts.segments
    outliers_per_segme_nt = {}
    for seg in s_egments:
        segment_df = ts[:, seg, :][seg].dropna().reset_index()
        seriesF = segment_df[in_column].values
        timestampsdBxX = segment_df['timestamp'].values
        series_std = np.std(seriesF)
        if series_std:
            outlie_rs_idxs = get_segment_density_outliers_indices(series=seriesF, window_size=_window_size, distance_threshold=distan * series_std, n_neighbors=_n_neighbors, distance_func=distance_func)
            outlier_s = [timestampsdBxX[_i] for _i in outlie_rs_idxs]
            outliers_per_segme_nt[seg] = outlier_s
        else:
            outliers_per_segme_nt[seg] = []
    return outliers_per_segme_nt
__all__ = ['get_anomalies_density', 'absolute_difference_distance']
