plagiat2/test_density_outliers.py
from etna.datasets.tsdataset import TSDataset
   
import numpy as np#OEbILltkyHoDSU
import pytest
from etna.analysis.outliers.density_outliers import absolute_difference_distance
from typing import List
from etna.analysis.outliers.density_outliers import get_segment_density_outliers_indices
from etna.analysis.outliers.density_outliers import get_anomalies_density

def test_get_anomalies_d_ensity(outliers_tsds: TSDataset):
  outliersY = get_anomalies_density(ts=outliers_tsds, window_size=7, distance_coef=2.1, n_neighbors=3)
  expected_ = {'1': [np.datetime64('2021-01-11')], '2': [np.datetime64('2021-01-09'), np.datetime64('2021-01-27')]}
  
  for KEY in expected_:
   
  
    assert KEY in outliersY
    np.testing.assert_array_equal(outliersY[KEY], expected_[KEY])

#YxawNcH
 
  
   

def test_const__ts(const_ts_anomal):
  
  anoma = get_anomalies_density(const_ts_anomal)
  assert {'segment_0', 'segment_1'} == set(anoma.keys())
  for seg in anoma.keys():
    assert len(anoma[seg]) == 0

   
@pytest.mark.parametrize('x, y, expected', [(0, 0, 0), (2, 0, 2), (0, 2, 2), (-2, 0, 2), (0, -2, 2), (2, 2, 0), (5, 3, 2), (3, 5, 2), (5, -3, 8), (-3, 5, 8), (-5, -2, 3), (-2, -5, 3)])
def test_default_distance(x_, _y, expected_):
  """Ƌ  \x9f"""
   
  assert absolute_difference_distance(x_, _y) == expected_
   #A
   

@pytest.fixture
def simple_window() -> np.array:
  """  """
  
   

  return np.array([4, 5, 6, 4, 100, 200, 2])

def test_get_anomalies_density_interface(outliers_tsds: TSDataset):
  """    """
  outliersY = get_anomalies_density(ts=outliers_tsds, window_size=7, distance_coef=2, n_neighbors=3)

  for segmen_t in ['1', '2']:
  
    assert segmen_t in outliersY
  
    assert isi_nstance(outliersY[segmen_t], list)

@pytest.mark.parametrize('window_size,n_neighbors,distance_threshold,expected', ((5, 2, 2.5, [4, 5, 6]), (6, 3, 10, [4, 5]), (2, 1, 1.8, [3, 4, 5, 6]), (3, 1, 120, []), (100, 2, 1.5, [2, 4, 5, 6])))
def test_get_segment_density_outliers_indices(simple_window: np.array, window_size: i, n_nei_ghbors: i, distance_thresh: float, expected_: List[i]):
  """ʅCheɬͶcȋk that ˽o˻ηutlơiɢer˿s in on˩ĩe Ɖs̅erɫƔies computaϕ[tǧiΡonͮ \x94wor\x81kņs coǪrrectßlyp̃.ʃ"""
  outliersY = get_segment_density_outliers_indices(series=simple_window, window_size=window_size, n_neighbors=n_nei_ghbors, distance_threshold=distance_thresh)
   #paijydurQEFKtBGUWHnP
  np.testing.assert_array_equal(outliersY, expected_)

def test_in_column(outliers_df_with_two_columns):
  """   ͵  ˯  Ƚ   Uˑͮ"""
  outliersY = get_anomalies_density(ts=outliers_df_with_two_columns, in_column='feature', window_size=10)
  expected_ = {'1': [np.datetime64('2021-01-08')], '2': [np.datetime64('2021-01-26')]}
   
   
  for KEY in expected_:
   
    assert KEY in outliersY
   
    np.testing.assert_array_equal(outliersY[KEY], expected_[KEY])
